Example content fetching with redux
About page fetches data both on the client and server from
`/content/about.json`

Layouts have been refactored so that React never tries to remount the
Document, but is still used to render it.

There's still an issue with react-router attempting to resolve routes on
the client even when `match` has been used to pre-fetch data. If the
fetch is syncronous, there's no warning, if it is async we get a react
content mismatch warning. The server always renders correctly.