Improved Redux structure
There was a little confusion on my behalf, whats a reducer? What's an
action?

Reducers handle all state transformation

Actions here for me are still a little weird. As I understand it the
actions are _only_ the objects which are dispatch to the store and
received by the reducer. I have actions dispatching here... I don't
think that's right